// Front/ProEventos-App/src/app/models/identity/UserUpdate.ts
export class UserUpdate {
  titulo: string | undefined;
  funcao: string | undefined;
  userName: string | undefined;
  primeiroNome: string | undefined;
  ultimoNome: string | undefined;
  email: string | undefined;
  phoneNumber: string | undefined;
  descricao: string | undefined;
  password: string | undefined;
  token: string | undefined;
}